Applied suggestions by jfg

+++ tools/release/lpsreach/lpsreach.cpp
@@ -32,7 +32,7 @@ class lpsreach_tool: public parallel_tool<rewriter_tool<input_output_tool>>
 
     // Lace options
     std::size_t lace_n_workers = 1;
-    std::size_t lace_dqsize = static_cast<long>(1024*1024)*4; // set large default
+    std::size_t lace_dqsize = static_cast<std::size_t>(1024*1024)*4; // set large default
     std::size_t lace_stacksize = 0; // use default
 
     // Sylvan options
